Fix warning 4.9.2

+++ src/Space/Shape/HyperCube.hpp
@@ -214,20 +214,20 @@ public:
 		size_t lin_id = 0;
 
 		// Cumulative value
-		size_t val = 0;
-		size_t cum_val = 0;
+		long int val = 0;
+		long int cum_val = 0;
 		for(long int i = d - 1; i >= 0 ; i--)
 		{
 			// check the out-of-bound, outside is assumed to be -1 so  (- pos_n_zero[i+1] - 1) = 0
-			if (i+1 < d)
+			if (i+1 < (long int)d)
 				val = pos_n_zero[i] - pos_n_zero[i+1] - 1;
 			else
 				val = pos_n_zero[i];
 
-			for (long int j = 0 ; j < val; j++)
+			for (long int j = 0 ; j < (long int)val; j++)
 			{
 				// C is not safe check the limit
-				if (((long int)dim)-cum_val-j-1 >= 0 && i > 0 && dim-cum_val-j >= i)
+				if (((long int)dim)-cum_val-j-1 >= 0 && i > 0 && ((long int)dim)-cum_val-j >= i)
 					lin_id += openfpm::math::C(dim-cum_val-j-1,i);
 				else
 					lin_id += 1;
